+++ src/components/settings/ContractVersion/index.tsx
@@ -1,5 +1,5 @@
-import { useMemo } from 'react'
-import { Box, SvgIcon, Typography, Alert, AlertTitle, Skeleton } from '@mui/material'
+import { useContext, useMemo } from 'react'
+import { Box, SvgIcon, Typography, Alert, AlertTitle, Skeleton, Button } from '@mui/material'
 import { ImplementationVersionState } from '@safe-global/safe-gateway-typescript-sdk'
 import { LATEST_SAFE_VERSION } from '@/config/constants'
 import { sameAddress } from '@/utils/addresses'
@@ -8,10 +8,13 @@ import { MasterCopyDeployer, useMasterCopies } from '@/hooks/useMasterCopies'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import CheckCircleIcon from '@mui/icons-material/CheckCircle'
 import InfoIcon from '@/public/images/notifications/info.svg'
-
-import UpdateSafeDialog from './UpdateSafeDialog'
+import { TxModalContext } from '@/components/tx-flow'
+import UpdateSafeFlow from '@/components/tx-flow/flows/UpdateSafe'
 import ExternalLink from '@/components/common/ExternalLink'
+import CheckWallet from '@/components/common/CheckWallet'
+
 export const ContractVersion = () => {
+  const { setTxFlow } = useContext(TxModalContext)
   const [masterCopies] = useMasterCopies()
   const { safe, safeLoaded } = useSafeInfo()
   const masterCopyAddress = safe.implementation.value
@@ -40,12 +43,20 @@ export const ContractVersion = () => {
               icon={<SvgIcon component={InfoIcon} inheritViewBox color="secondary" />}
             >
               <AlertTitle sx={{ fontWeight: 700 }}>New version is available: {LATEST_SAFE_VERSION}</AlertTitle>
+
               <Typography mb={3}>
                 Update now to take advantage of new features and the highest security standards available. You will need
                 to confirm this update just like any other transaction.{' '}
                 <ExternalLink href={safeMasterCopy?.deployerRepoUrl}>GitHub</ExternalLink>
               </Typography>
-              <UpdateSafeDialog />
+
+              <CheckWallet>
+                {(isOk) => (
+                  <Button onClick={() => setTxFlow(<UpdateSafeFlow />)} variant="contained" disabled={!isOk}>
+                    Update
+                  </Button>
+                )}
+              </CheckWallet>
             </Alert>
           ) : (
             <Typography display="flex" alignItems="center">
