+++ src/components/tx-flow/flows/UpdateSafe/index.tsx
@@ -1,0 +1,12 @@
+import TxLayout from '@/components/tx-flow/common/TxLayout'
+import { UpdateSafeReview } from './UpdateSafeReview'
+
+const UpdateSafeFlow = () => {
+  return (
+    <TxLayout title="Update Safe Account version">
+      <UpdateSafeReview />
+    </TxLayout>
+  )
+}
+
+export default UpdateSafeFlow
